chore(tools): add "start_script.sh" and "stop_script.sh" file

+++ script/deploy/stop_script.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ ! -f exporter.pid ]; then
+    echo "The SNMP exporter is not started."
+    exit 0
+fi
+
 kill -9 $(cat exporter.pid)
 
 rm -rf exporter.pid
